requirements.txt
absl-py==0.8.1
asn1crypto==0.24.0
astor==0.8.0
certifi==2021.5.30
cffi==1.15.0
cryptography==2.1.4
cycler==0.10.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.24.3
h5py==2.10.0
idna==2.6
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
Markdown==3.1.1
matplotlib==3.3.4
numpy==1.19.5
opencv-python-headless==4.5.1.48
opt-einsum==3.1.0
Pillow==8.1.0
protobuf==3.19.4
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
scipy==1.5.4
SecretStorage==2.3.1
six==1.11.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
Werkzeug==0.16.0
wrapt==1.11.2
